requirements.txt
textual
rich
psycopg[binary]
faker
PyYAML
requests
networkx
